fix: improve ergonomics when referring to functions from quoted code
Previously, we would treat a function call like `xyz(...)` as `xyz` being a free variable. This is correct if `xyz` is another piece of quoted code, but more frequently in user code `xyz` is just some external function. This switches the default to assuming that it is an external function, but when using captured functions `let a = xyz;` can be used.