feat: make code-completion work with Rust Analyzer (#9)
Rust Analyzer seems to trip up on the input to the `q!(...)` macro being double-emitted, so we use environment variables to detect when we are being called from Rust Analyzer and work around this issue. The end result is that when code completing from syntactically valid code, the completions work!

Under the hood, this also tweaks the free variable analysis. Before, each free variable would be replaced in the body with an `__free` variant. Now, we use the same identifier as the original variable and use shadowing rules to introduce the unquoted value. This should reduce brittleness when the macro fails to identify a free variable that's used somewhere else.