feat(stageleft)!: remove `stageleft_devel`, make macro entrypoints opt-in via a feature
Previously, we defaulted to assume that there may be macro entrypoints, and so would always compile an updated `__staged` module whenever building any crate with quoted code. This flips the default, so we do not compile `__staged` by default, ans instead only do it if the compiler is invoked with `--cfg stageleft_trybuild` (for trybuild-style compilation in Hydro), or with the `stageleft_macro_entrypoint` feature enabled (for macro entrypoints).

The downstream impact of this is that in workspaces that do not have any macro entrypoints (such as Hydro), we skip compiling `__staged` in stage 0. This also dramatically reduces thrashing in Rust Analyzer, since we are not re-generating new Rust sources on each edit.

As part of the rewrite, we also simplify the logic of rewriting types and paths to use the `__staged` module. Now, all crates with quoted code register themselves in a constructor, so that we know to rewrite any references to types inside them.

This is a breaking change, since it flips the defaults for macro entrypoints.