feat(stageleft_tool)!: use features to gate test-mode staged code
Previously, we would generate either the `__staged` module with `#[cfg(test)]` APIs included or not based on a boolean flag. This makes it harder to share the generated staged code across separate deployments, since some may be in test mode or not. With this PR, we instead take in a string which is the feature under which test APIs should be included. This allows us to generate the `__staged` module once and share it across several targets.